Version 0.02

+++ src/data.hpp
@@ -1,15 +1,15 @@
 #pragma once
 
 #include"zxlb.hpp"
-#include"vec3.hpp"
+#include"vec_s.hpp"
 
 struct SynapseData {
     str id;
     std::size_t data_size;
     vec<i64> time_indexes;
     vec<double> input;
-    vec<i64> parent_id;
-    vec<vec<i64>> children_ids;
+    vec<int> parent_id;
+    vec<vec<int>> children_ids;
     vec<VecS> locations;
     vec<double> strength;
     vec<i64> spikes;
